download tor works on macos
